refactor(layer): 계층화 구조 적용한 패키지 재분리
변경내역
- presentation 패키지에 컨트롤러와 뷰 포함
- domain 패키지에 모델과 dto 포함
- utils 패키지명을 common 으로 변경

기존 mvc 패턴을 그대로 사용하던 패키지 분리에서 계층화 구조 적용
