+++ src/main/java/racingcar/racing/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar.racing.model;
+package racingcar.racing.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
