+++ src/test/java/racingcar/racing/domain/CarTest.java
@@ -1,0 +1,4 @@
+package racingcar.racing.domain;
+
+public class CarTest {
+}
