+++ src/test/java/racingcar/racing/domain/RacingGameTest.java
@@ -1,0 +1,5 @@
+package racingcar.racing.domain;
+
+class RacingGameTest {
+
+}
