--- src/test/java/racingcar/racing/model/CarTest.java
@@ -1,4 +1,0 @@
-package racingcar.racing.model;
-
-public class CarTest {
-}
