--- src/test/java/racingcar/racing/model/RacingGameTest.java
@@ -1,5 +1,0 @@
-package racingcar.racing.model;
-
-class RacingGameTest {
-
-}
